Version 1.7.6.2
Changelog:
- Saving changes before migrating to null safety.

+++ flutter_app_devotion_sim/lib/classes/stats.dart
@@ -1,11 +1,11 @@
 class Stats {
   final String time;
-  final double gas;
-  final double speed;
-  final double gear;
-  final double lean;
-  final double frontBrake;
-  final double rearBrake;
+  final int gas;
+  final int speed;
+  final int gear;
+  final int lean;
+  final int frontBrake;
+  final int rearBrake;
 
   Stats({
     this.time,
@@ -20,12 +20,12 @@ class Stats {
   factory Stats.fromJson(Map<String, dynamic> json) {
     return Stats(
         time: json['Time'] as String,
-        gas: json['Gas'] as double,
-        speed: json['Speed'] as double,
-        gear: json['Gear'] as double,
-        lean: json['Lean'] as double,
-        frontBrake: json['Front Brake'] as double,
-        rearBrake: json['Rear Brake'] as double
+        gas: json['Gas'] as int,
+        speed: json['Speed'] as int,
+        gear: json['Gear'] as int,
+        lean: json['Lean'] as int,
+        frontBrake: json['Front Brake'] as int,
+        rearBrake: json['Rear Brake'] as int
     );
   }
 }
